// ClientApp/generated/ProjectFieldTypeModel.ts
/*  GENERTAED automatically using Models.tst. DO NOT CHANGE HERE */

//Generated from class commonInterfaces.dbDataTypes.ProjectFieldTypeModel
export enum ProjectFieldTypeModel {
  text = 'text',
  number = 'number',
  date = 'date',
  auto_inc = 'auto_inc',
  timestamp = 'timestamp',
  user_list = 'user_list',
  boolean = 'boolean',
  group = 'group',
  currency = 'currency',
  externalIntegration = 'externalIntegration',
}